version 1.0.5 code added

+++ frontend/src/Components/SalePage/Sales.jsx
@@ -5,51 +5,33 @@ const Sale = () => {
   const naviagate = useNavigate();
 
   const [productDetails, setProductDetails] = useState({
-    brand: "",
-    product: "",
-    category: "",
-    size: "",
-    mrp: "",
-    quantitySold: 1,
+    code: "",
     sellingPrice: "",
+    quantitySold: 1,
   });
 
-  const [dropdownOptions, setDropdownOptions] = useState({
-    products: [],
-    categories: [],
-    sizes: [],
-  });
-
-  const [productOfSelectedBrand, setProductOfSelectedBrand] = useState([]);
-  const [availableQuantity, setAvailableQuantity] = useState(0);
-  const [isFormValid, setIsFormValid] = useState(false);
+  const [buttonActive, setButtonActive] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
   const [showTooltip, setShowTooltip] = useState(false);
-  const [mrpList, setMRPList] = useState([]);
+  const [availableQuantity, setAvailableQuantity] = useState(0);
   const [fetchingQuantity, setFetchingQuantity] = useState(false);
 
   useEffect(() => {
-    function isUserLogedIn() {
+    function isUserLoggedIn() {
       if (!window.localStorage.getItem("userInfo")) {
         naviagate("/login");
       }
     }
-    isUserLogedIn();
+    isUserLoggedIn();
   }, [naviagate]);
 
   // Fetch available quantity when brand, product, category, mrp, or size changes
   const fetchAvailableQuantity = async () => {
     try {
-      if (
-        productDetails.brand &&
-        productDetails.product &&
-        productDetails.category &&
-        productDetails.size &&
-        productDetails.mrp
-      ) {
+      if (productDetails.code) {
         setFetchingQuantity(true);
         const response = await fetch(
-          `/api/availablequantity?brand=${productDetails.brand}&product=${productDetails.product}&category=${productDetails.category}&size=${productDetails.size}&mrp=${productDetails.mrp}`,
+          `/api/availablequantity?code=${productDetails.code}`,
           {
             headers: {
               Authorization: `Bearer ${
@@ -61,11 +43,17 @@ const Sale = () => {
         if (response.ok) {
           var { availableQuantity } = await response.json();
           setAvailableQuantity(availableQuantity);
+          if (availableQuantity > 0) {
+            setButtonActive(true);
+          }
         } else {
+          setButtonActive(false);
           setAvailableQuantity(0);
           console.log("Failed to fetch available quantity");
-          window.localStorage.clear();
-          naviagate("/login");
+          if (response.status === 401) {
+            window.localStorage.clear();
+            naviagate("/login");
+          }
         }
         setFetchingQuantity(false);
       } else {
@@ -75,164 +63,29 @@ const Sale = () => {
       console.log(error);
     }
   };
-  // Fetch mrp list when brand, product, category, or size changes
-  const fetchMrpList = async () => {
-    try {
-      if (
-        productDetails.brand &&
-        productDetails.product &&
-        productDetails.category &&
-        productDetails.size
-      ) {
-        const response = await fetch(
-          `/api/item/list?brand=${productDetails.brand}&product=${productDetails.product}&category=${productDetails.category}&size=${productDetails.size}`,
-          {
-            headers: {
-              Authorization: `Bearer ${
-                JSON.parse(window.localStorage.getItem("userInfo")).token
-              }`,
-            },
-          }
-        );
-
-        if (response.ok) {
-          const listOfItems = await response.json();
-
-          // Extract unique MRPs from the list of items
-          const uniqueMRPs = Array.from(
-            new Set(listOfItems.map((item) => item.mrp))
-          );
-
-          // Set the MRP list state
-          setMRPList(uniqueMRPs);
-        } else {
-          setMRPList([]);
-          console.log("Failed to fetch list");
-          window.localStorage.clear();
-          naviagate("/login");
-        }
-      } else {
-        setMRPList([]);
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  // Fetch dropdown options from the server
-  const fetchDropdownOptions = async () => {
-    if (window.localStorage.getItem("userInfo")) {
-      try {
-        const response = await fetch("/api/dropdownoption/dropdownoptions", {
-          headers: {
-            Authorization: `Bearer ${
-              JSON.parse(window.localStorage.getItem("userInfo")).token
-            }`,
-          },
-        });
-        if (response.ok) {
-          const options = await response.json();
-          setDropdownOptions(options);
-        } else {
-          console.error("Failed to fetch dropdown options");
-          window.localStorage.clear();
-          naviagate("/login");
-        }
-      } catch (error) {
-        console.log(error);
-      }
-    }
-  };
-
-  // form validation
-  const checkFormValidity = ({
-    brand,
-    product,
-    category,
-    size,
-    quantitySold,
-    sellingPrice,
-  }) =>
-    brand !== "" &&
-    product !== "" &&
-    category !== "" &&
-    size !== "" &&
-    availableQuantity !== "" &&
-    parseInt(quantitySold) === 1 &&
-    parseInt(sellingPrice) > 0;
-
-  useEffect(() => {
-    fetchDropdownOptions();
-    // eslint-disable-next-line
-  }, []);
-
   useEffect(() => {
     fetchAvailableQuantity();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [
-    productDetails.brand,
-    productDetails.product,
-    productDetails.category,
-    productDetails.size,
-    productDetails.mrp,
-  ]);
-  useEffect(() => {
-    fetchMrpList();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [
-    productDetails.brand,
-    productDetails.product,
-    productDetails.category,
-    productDetails.size,
-  ]);
-
-  useEffect(() => {
-    setIsFormValid(checkFormValidity(productDetails));
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [productDetails, availableQuantity]);
+    // eslint-disable-next-line
+  }, [productDetails.code]);
 
   // handle changes in inputs
   const handleInputChange = (e) => {
     const { name, value } = e.target;
 
     // Ensure quantitySold and sellingPrice are positive values
-    if (
-      (name === "quantitySold" || name === "sellingPrice") &&
-      parseFloat(value) < 0
-    ) {
+    if (name === "sellingPrice" && parseFloat(value) < 0) {
       return;
     }
 
     setProductDetails((prevDetails) => ({
       ...prevDetails,
-      [name]: value,
-    }));
-  };
-
-  const handleBrandChange = (e) => {
-    const { name, value } = e.target;
-    var selectedBrandProducts = dropdownOptions.products.filter(
-      (product) => product.brand === value
-    );
-    selectedBrandProducts = selectedBrandProducts[0]?.products || [];
-    setProductOfSelectedBrand(selectedBrandProducts);
-    setProductDetails((prevDetails) => ({
-      ...prevDetails,
-      [name]: value,
+      [name]: value.toUpperCase(),
     }));
   };
 
   const handleSale = async () => {
-    // Validation: Check if all fields are filled
-    if (!isFormValid) {
-      console.error(
-        "Please fill in all fields and ensure the quantity sold and selling price are positive values."
-      );
-      return;
-    }
-
     setIsLoading(true);
-
+    setButtonActive(false);
     try {
       const response = await fetch("/api/sell", {
         method: "POST",
@@ -248,12 +101,9 @@ const Sale = () => {
       if (response.ok) {
         // Reset the form after successful sale
         setProductDetails({
-          brand: "",
-          product: "",
-          category: "",
-          size: "",
-          quantitySold: "",
+          code: "",
           sellingPrice: "",
+          quantitySold: 1,
         });
         console.log("Product sold from inventory successfully!");
         setShowTooltip(true);
@@ -276,152 +126,23 @@ const Sale = () => {
   return (
     <div className="bg-white p-6 rounded shadow-lg">
       <h2 className="text-2xl font-bold mb-6">Sale</h2>
-      <div className="mb-4">
-        <label
-          htmlFor="brand"
-          className="block text-sm font-medium text-gray-700"
-        >
-          Brand
-        </label>
-        <select
-          name="brand"
-          value={productDetails.brand}
-          onChange={handleBrandChange}
-          className="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-        >
-          <option value="" disabled>
-            Select Brand
-          </option>
-          {dropdownOptions.products.map((product) => (
-            <option key={product._id} value={product.brand}>
-              {product.brand}
-            </option>
-          ))}
-        </select>
-      </div>
 
       <div className="mb-4">
         <label
-          htmlFor="product"
+          htmlFor="code"
           className="block text-sm font-medium text-gray-700"
         >
-          Product
-        </label>
-        <select
-          name="product"
-          value={productDetails.product}
-          onChange={handleInputChange}
-          className="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-        >
-          <option value="" disabled>
-            Select Product
-          </option>
-          {productOfSelectedBrand.map((product) => (
-            <option key={product} value={product}>
-              {product}
-            </option>
-          ))}
-        </select>
-      </div>
-
-      <div className="mb-4">
-        <label
-          htmlFor="category"
-          className="block text-sm font-medium text-gray-700"
-        >
-          Sub-category
-        </label>
-        <select
-          name="category"
-          value={productDetails.category}
-          onChange={handleInputChange}
-          className="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-        >
-          <option value="" disabled>
-            Select Sub-category
-          </option>
-          {dropdownOptions.categories.map((categoryObj) =>
-            categoryObj.category.map((subcategory) => (
-              <option key={subcategory} value={subcategory}>
-                {subcategory}
-              </option>
-            ))
-          )}
-        </select>
-      </div>
-
-      <div className="mb-4">
-        <label
-          htmlFor="size"
-          className="block text-sm font-medium text-gray-700"
-        >
-          Size
-        </label>
-        <select
-          name="size"
-          value={productDetails.size}
-          onChange={handleInputChange}
-          className="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-        >
-          <option value="" disabled>
-            Select Size
-          </option>
-          {dropdownOptions.sizes.map((sizeObj) =>
-            sizeObj.size.map((size) => (
-              <option key={size} value={size}>
-                {size}
-              </option>
-            ))
-          )}
-        </select>
-      </div>
-
-      <div className="mb-4">
-        <label
-          htmlFor="mrp"
-          className="block text-sm font-medium text-gray-700"
-        >
-          MRP
-        </label>
-        <select
-          name="mrp"
-          value={productDetails.mrp}
-          onChange={handleInputChange}
-          className="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-        >
-          <option value="" disabled>
-            Select MRP
-          </option>
-          {mrpList.map((mrp) => (
-            <option key={mrp} value={mrp}>
-              {mrp}
-            </option>
-          ))}
-        </select>
-      </div>
-
-      <p className="mb-4 text-green-500">
-        Available Quantity:{" "}
-        {fetchingQuantity ? "Fetching Quantity..." : availableQuantity}
-      </p>
-
-      <div className="mb-4">
-        <label
-          htmlFor="quantitySold"
-          className="block text-sm font-medium text-gray-700"
-        >
-          Quantity Sold
+          Code
         </label>
         <input
-          type="number"
-          placeholder="Quantity Sold"
-          name="quantitySold"
-          value={productDetails.quantitySold}
-          // onChange={handleInputChange}
+          name="code"
+          type="string"
+          placeholder="Product code"
+          value={productDetails.code}
+          onChange={handleInputChange}
           className="mt-1 block w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-        />
+        ></input>
       </div>
-
       <div className="mb-4">
         <label
           htmlFor="sellingPrice"
@@ -439,17 +160,45 @@ const Sale = () => {
         />
       </div>
 
+      <p className="mb-4 text-green-500">
+        {/* Update with your available quantity logic */}
+        Available Quantity:{" "}
+        {fetchingQuantity ? "Fetching Quantity..." : availableQuantity}
+      </p>
+
       {showTooltip && (
-        <div className="text-green-500 text-sm mb-4">
-          Product sold from inventory successfully!
-        </div>
+        <>
+          <div
+            className="fixed inset-0 flex items-center justify-center"
+            style={{ backgroundColor: "rgba(0, 0, 0, 0.5)", zIndex: 999 }}
+          >
+            <div
+              className="flex items-center justify-center bg-green-100 border border-green-400 text-green-700 px-4 py-3 rounded relative w-full m-5"
+              role="alert"
+            >
+              <strong className="font-bold">Product sold</strong>
+              <span className="absolute top-0 bottom-0 right-0 px-4 py-3">
+                <svg
+                  className="fill-current h-6 w-6 text-green-500 cursor-pointer"
+                  role="button"
+                  xmlns="http://www.w3.org/2000/svg"
+                  viewBox="0 0 20 20"
+                  onClick={() => setShowTooltip(false)}
+                >
+                  <title>Close</title>
+                  <path d="M14.348 14.849a1.2 1.2 0 0 1-1.697 0L10 11.819l-2.651 3.029a1.2 1.2 0 1 1-1.697-1.697l2.758-3.15-2.759-3.152a1.2 1.2 0 1 1 1.697-1.697L10 8.183l2.651-3.031a1.2 1.2 0 1 1 1.697 1.697l-2.758 3.152 2.758 3.15a1.2 1.2 0 0 1 0 1.698z" />
+                </svg>
+              </span>
+            </div>
+          </div>
+        </>
       )}
 
       <button
         onClick={handleSale}
-        disabled={!isFormValid || isLoading}
+        disabled={!buttonActive || isLoading}
         className={`w-full py-2 px-4 rounded focus:outline-none focus:shadow-outline-indigo active:bg-indigo-800 ${
-          isFormValid
+          buttonActive
             ? "bg-indigo-500 text-white"
             : "bg-gray-300 text-black opacity-50 cursor-not-allowed"
         }`}
